ci: remove build matrix
We want to fail the build if the test coverage of the pull request is not 100%. To archive that goal we need to combine coverage from different tox environments. To do this the coverage data from all tox environments should be available on the same machine. To do this we need to stop using the build matrix on our CI. Yes, it will increase the build time since there is no parallel execution of the environments anymore. Yes, there are some advanced alternatives to this approach which will not increase the build time. We decided to take the simples option until build time became an ACTUAL problem.